Modified plugins to be based on folders instead of individual files. Made this change due to the fact that the slack plugin got so large and complicated that I needed to break it into individual files
